Split LoginButtons component
Fix SSR issues
